moved test util method to DataStreamTestHelper
