+++ test/framework/src/main/java/org/elasticsearch/cluster/DataStreamTestHelper.java
@@ -16,14 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 package org.elasticsearch.cluster;
 
 import org.apache.lucene.util.LuceneTestCase;
 import org.elasticsearch.Version;
 import org.elasticsearch.cluster.metadata.DataStream;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
+import org.elasticsearch.cluster.metadata.Metadata;
 import org.elasticsearch.common.UUIDs;
+import org.elasticsearch.common.collect.Tuple;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.test.ESTestCase;
@@ -31,6 +32,7 @@ import org.elasticsearch.test.ESTestCase;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
+import java.util.stream.Collectors;
 
 import static org.elasticsearch.cluster.metadata.DataStream.getDefaultBackingIndexName;
 import static org.elasticsearch.cluster.metadata.IndexMetadata.SETTING_INDEX_UUID;
@@ -101,5 +103,48 @@ public final class DataStreamTestHelper {
         String dataStreamName = ESTestCase.randomAlphaOfLength(10).toLowerCase(Locale.ROOT);
         indices.add(new Index(getDefaultBackingIndexName(dataStreamName, generation), UUIDs.randomBase64UUID(LuceneTestCase.random())));
         return new DataStream(dataStreamName, createTimestampField("@timestamp"), indices, generation);
+    }
+
+    /**
+     * Constructs {@code ClusterState} with the specified data streams and indices.
+     *
+     * @param dataStreams The names of the data streams to create with their respective number of backing indices
+     * @param indexNames  The names of indices to create that do not back any data streams
+     */
+    public static ClusterState getClusterStateWithDataStreams(List<Tuple<String, Integer>> dataStreams, List<String> indexNames) {
+        Metadata.Builder builder = Metadata.builder();
+
+        List<IndexMetadata> allIndices = new ArrayList<>();
+        for (Tuple<String, Integer> dsTuple : dataStreams) {
+            List<IndexMetadata> backingIndices = new ArrayList<>();
+            for (int backingIndexNumber = 1; backingIndexNumber <= dsTuple.v2(); backingIndexNumber++) {
+                backingIndices.add(createIndexMetadata(getDefaultBackingIndexName(dsTuple.v1(), backingIndexNumber), true));
+            }
+            allIndices.addAll(backingIndices);
+
+            DataStream ds = new DataStream(
+                dsTuple.v1(),
+                createTimestampField("@timestamp"),
+                backingIndices.stream().map(IndexMetadata::getIndex).collect(Collectors.toList()),
+                dsTuple.v2()
+            );
+            builder.put(ds);
+        }
+
+        for (String indexName : indexNames) {
+            allIndices.add(createIndexMetadata(indexName, false));
+        }
+
+        for (IndexMetadata index : allIndices) {
+            builder.put(index, false);
+        }
+
+        return ClusterState.builder(new ClusterName("_name")).metadata(builder).build();
+    }
+
+    private static IndexMetadata createIndexMetadata(String name, boolean hidden) {
+        Settings.Builder b = Settings.builder().put(IndexMetadata.SETTING_VERSION_CREATED, Version.CURRENT).put("index.hidden", hidden);
+
+        return IndexMetadata.builder(name).settings(b).numberOfShards(1).numberOfReplicas(1).build();
     }
 }
