+++ x-pack/plugin/data-streams/src/test/java/org/elasticsearch/xpack/datastreams/action/DeleteDataStreamTransportActionTests.java
@@ -6,18 +6,15 @@
 
 package org.elasticsearch.xpack.datastreams.action;
 
-import org.elasticsearch.Version;
-import org.elasticsearch.cluster.ClusterName;
 import org.elasticsearch.cluster.ClusterState;
+import org.elasticsearch.cluster.DataStreamTestHelper;
 import org.elasticsearch.cluster.SnapshotsInProgress;
 import org.elasticsearch.cluster.metadata.DataStream;
-import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.cluster.metadata.Metadata;
 import org.elasticsearch.cluster.metadata.MetadataDeleteIndexService;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.collect.ImmutableOpenMap;
 import org.elasticsearch.common.collect.Tuple;
-import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.snapshots.Snapshot;
 import org.elasticsearch.snapshots.SnapshotId;
@@ -25,13 +22,10 @@ import org.elasticsearch.snapshots.SnapshotInProgressException;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.xpack.core.action.DeleteDataStreamAction;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
-import static org.elasticsearch.cluster.DataStreamTestHelper.createTimestampField;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Matchers.any;
@@ -44,7 +38,7 @@ public class DeleteDataStreamTransportActionTests extends ESTestCase {
         final String dataStreamName = "my-data-stream";
         final List<String> otherIndices = randomSubsetOf(List.of("foo", "bar", "baz"));
 
-        ClusterState cs = getClusterStateWithDataStreams(List.of(new Tuple<>(dataStreamName, 2)), otherIndices);
+        ClusterState cs = DataStreamTestHelper.getClusterStateWithDataStreams(List.of(new Tuple<>(dataStreamName, 2)), otherIndices);
         DeleteDataStreamAction.Request req = new DeleteDataStreamAction.Request(new String[] { dataStreamName });
         ClusterState newState = DeleteDataStreamTransportAction.removeDataStream(getMetadataDeleteIndexService(), cs, req);
         assertThat(newState.metadata().dataStreams().size(), equalTo(0));
@@ -56,7 +50,7 @@ public class DeleteDataStreamTransportActionTests extends ESTestCase {
 
     public void testDeleteMultipleDataStreams() {
         String[] dataStreamNames = { "foo", "bar", "baz", "eggplant" };
-        ClusterState cs = getClusterStateWithDataStreams(
+        ClusterState cs = DataStreamTestHelper.getClusterStateWithDataStreams(
             List.of(
                 new Tuple<>(dataStreamNames[0], randomIntBetween(1, 3)),
                 new Tuple<>(dataStreamNames[1], randomIntBetween(1, 3)),
@@ -82,7 +76,7 @@ public class DeleteDataStreamTransportActionTests extends ESTestCase {
         final String dataStreamName2 = "my-data-stream2";
         final List<String> otherIndices = randomSubsetOf(List.of("foo", "bar", "baz"));
 
-        ClusterState cs = getClusterStateWithDataStreams(
+        ClusterState cs = DataStreamTestHelper.getClusterStateWithDataStreams(
             List.of(new Tuple<>(dataStreamName, 2), new Tuple<>(dataStreamName2, 2)),
             otherIndices
         );
@@ -126,7 +120,7 @@ public class DeleteDataStreamTransportActionTests extends ESTestCase {
     public void testDeleteNonexistentDataStream() {
         final String dataStreamName = "my-data-stream";
         String[] dataStreamNames = { "foo", "bar", "baz", "eggplant" };
-        ClusterState cs = getClusterStateWithDataStreams(
+        ClusterState cs = DataStreamTestHelper.getClusterStateWithDataStreams(
             List.of(
                 new Tuple<>(dataStreamNames[0], randomIntBetween(1, 3)),
                 new Tuple<>(dataStreamNames[1], randomIntBetween(1, 3)),
@@ -160,49 +154,6 @@ public class DeleteDataStreamTransportActionTests extends ESTestCase {
         });
 
         return s;
-    }
-
-    /**
-     * Constructs {@code ClusterState} with the specified data streams and indices.
-     *
-     * @param dataStreams The names of the data streams to create with their respective number of backing indices
-     * @param indexNames  The names of indices to create that do not back any data streams
-     */
-    public static ClusterState getClusterStateWithDataStreams(List<Tuple<String, Integer>> dataStreams, List<String> indexNames) {
-        Metadata.Builder builder = Metadata.builder();
-
-        List<IndexMetadata> allIndices = new ArrayList<>();
-        for (Tuple<String, Integer> dsTuple : dataStreams) {
-            List<IndexMetadata> backingIndices = new ArrayList<>();
-            for (int backingIndexNumber = 1; backingIndexNumber <= dsTuple.v2(); backingIndexNumber++) {
-                backingIndices.add(createIndexMetadata(DataStream.getDefaultBackingIndexName(dsTuple.v1(), backingIndexNumber), true));
-            }
-            allIndices.addAll(backingIndices);
-
-            DataStream ds = new DataStream(
-                dsTuple.v1(),
-                createTimestampField("@timestamp"),
-                backingIndices.stream().map(IndexMetadata::getIndex).collect(Collectors.toList()),
-                dsTuple.v2()
-            );
-            builder.put(ds);
-        }
-
-        for (String indexName : indexNames) {
-            allIndices.add(createIndexMetadata(indexName, false));
-        }
-
-        for (IndexMetadata index : allIndices) {
-            builder.put(index, false);
-        }
-
-        return ClusterState.builder(new ClusterName("_name")).metadata(builder).build();
-    }
-
-    private static IndexMetadata createIndexMetadata(String name, boolean hidden) {
-        Settings.Builder b = Settings.builder().put(IndexMetadata.SETTING_VERSION_CREATED, Version.CURRENT).put("index.hidden", hidden);
-
-        return IndexMetadata.builder(name).settings(b).numberOfShards(1).numberOfReplicas(1).build();
     }
 
 }
